+++ e2e/jest-css-transform-demo/src/Button.spec.tsx
@@ -1,3 +1,16 @@
+/**
+ * @jest-environment jsdom
+ */
+import '@testing-library/jest-dom/extend-expect';
+
+import { render } from '@testing-library/react';
+
 import Button from './Button';
 
 it('验证button组件存在', () => expect(Button).toBeDefined());
+
+it('按钮标题', () => {
+  const { getByTestId } = render(<Button data-testid="button">标题</Button>);
+
+  expect(getByTestId('button')).toHaveTextContent('标题');
+});