+++ e2e/jest-css-transform-demo/src/Button.tsx
@@ -1,9 +1,25 @@
-import React from 'react';
 import './Button.css';
 
 /**
- * 用于测试的按钮组件
+ * 按钮组件属性
  */
-export default function Button() {
-  return <div className="button">按钮</div>;
+interface ButtonProps {
+  /**
+   * 按钮中包含的元素
+   */
+  children?: React.ReactNode;
 }
+
+/**
+ * 用于测试的按钮组件
+ *
+ * @param props 属性
+ * @param props.children 子元素
+ */
+const Button: React.FC<ButtonProps> = ({ children, ...rest }) => (
+  <div className="button" {...rest}>
+    {children}
+  </div>
+);
+
+export default Button;